requirements.txt
pynacl
pythonp2p